rename to dashboard in additional-services
