+++ app/pages/sections/order-items/additional-services/dashboard/ui-tests/general.ui.test.js
@@ -33,7 +33,7 @@ const pageSetup = async (withAuth = true) => {
 
 const getLocation = ClientFunction(() => document.location.href);
 
-fixture('Additional-services page - general')
+fixture('Additional-services - Dashboard page - general')
   .page('http://localhost:1234/order/some-fake-page')
   .afterEach(async (t) => {
     const isDone = nock.isDone();
